pyproject.toml
[project]
name = "hardcover-mcp"
version = "0.1.2"
description = "MCP server for the Hardcover GraphQL API — library tracking and list management."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Kristian Edlund", email = "edlund@kristianedlund.dk" }
]
requires-python = ">=3.14"
keywords = ["mcp", "hardcover", "graphql", "books", "reading"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/kristianedlund/hardcover-mcp"
Repository = "https://github.com/kristianedlund/hardcover-mcp"
Issues = "https://github.com/kristianedlund/hardcover-mcp/issues"

[project.scripts]
hardcover-mcp = "hardcover_mcp.server:main"

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.11",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: tests that hit the real Hardcover API (require HARDCOVER_API_TOKEN)",
]

[tool.ruff]
target-version = "py314"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
